move host funcs
