+++ x/programs/wasmlanche/Cargo.toml
@@ -24,4 +24,3 @@ debug = ["std"]
 std = []
 test = ["std", "simulator"]
 simulator = ["std", "dep:simulator"]
-unit_tests = ["std"]
