+++ x/programs/wasmlanche/src/types.rs
@@ -25,7 +25,7 @@ impl From<Box<[u8]>> for ProgramId {
     }
 }
 
-#[cfg(feature = "unit_tests")]
+#[cfg(feature = "test")]
 impl ProgramId {
     #[must_use]
     pub fn new(bytes: Box<[u8]>) -> Self {
